Deduplicate task registration
When bridging a container into the universal model, we register domain
object added to the container as model object owned by the
project/settings. There is a way to notify when the known container
element is being registered via the model which wasn't implemented for
the TaskContainer.
